pixelcore/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ── Database (PostgreSQL via asyncpg, migrations via Alembic) ─────────────────
sqlalchemy==2.0.35
asyncpg==0.29.0          # async PostgreSQL driver (replaces aiosqlite)
alembic==1.13.3
greenlet==3.1.1

# ── Auth ───────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12

# ── Validation ─────────────────────────────────────────────────────────────────
pydantic[email]==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# ── Email ──────────────────────────────────────────────────────────────────────
sendgrid==6.11.0

# ── HTTP client (LLM API calls, OpenAI-compatible) ────────────────────────────
httpx==0.27.2

# ── Env ────────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── PDF resume parsing ─────────────────────────────────────────────────────────
pypdf==4.3.1

# ── Vision AI (optional — only when VISION_PROVIDER=deepface) ─────────────────
deepface==0.0.93
opencv-python-headless==4.10.0.84
tf-keras>=2.16.0
numpy>=1.24.0

# ── Speech-to-Text (local fallback when STT_PROVIDER=local) ───────────────────
faster-whisper==1.0.3

# ── WebSocket (pulled in by uvicorn[standard]; listed explicitly for clarity) ──
websockets>=12.0

# ── Task queue & cache ─────────────────────────────────────────────────────────
redis==5.0.8
celery[redis]==5.4.0
